native locks

+++ packages/server/src/routes/api/items.ts
@@ -5,16 +5,116 @@ import Router from '../../utils/Router';
 import { RouteType } from '../../utils/types';
 import { AppContext } from '../../utils/types';
 import * as fs from 'fs-extra';
-import { ErrorForbidden, ErrorMethodNotAllowed, ErrorNotFound, ErrorPayloadTooLarge, errorToPlainObject } from '../../utils/errors';
-import ItemModel, { ItemSaveOption, SaveFromRawContentItem } from '../../models/ItemModel';
+import { ErrorBadRequest, ErrorForbidden, ErrorMethodNotAllowed, ErrorNotFound, ErrorPayloadTooLarge, errorToPlainObject } from '../../utils/errors';
+import ItemModel, { ItemSaveOption, PaginatedItems, SaveFromRawContentItem } from '../../models/ItemModel';
 import { requestDeltaPagination, requestPagination } from '../../models/utils/pagination';
 import { AclAction } from '../../models/BaseModel';
 import { safeRemove } from '../../utils/fileUtils';
 import { formatBytes, MB } from '../../utils/bytes';
+import { Value } from '../../models/KeyValueModel';
 
 const router = new Router(RouteType.Api);
 
 const batchMaxSize = 1 * MB;
+
+interface LockHandlerResult {
+	handled: boolean;
+	response: any;
+}
+
+const lockHandler = async (path: SubPath, ctx: AppContext, requestBody: Buffer = null): Promise<LockHandlerResult | null> => {
+	// return { handled: false, response: null };
+	if (!path.id || !path.id.startsWith('root:/locks/')) return { handled: false, response: null };
+
+	const ownerId = ctx.joplin.owner.id;
+	const models = ctx.joplin.models;
+
+	const userKey = `locks::${ownerId}`;
+
+	// PUT /api/items/root:/locks/exclusive_cli_12cb74fa9de644958b2ccbc772cb4e29.json:/content
+
+	if (ctx.method === 'PUT') {
+		const itemName = models.item().pathToName(path.id);
+		const now = Date.now();
+
+		await models.keyValue().readThenWrite(userKey, async (value: Value) => {
+			const output = value ? JSON.parse(value as string) : {};
+			output[itemName] = {
+				name: itemName,
+				updated_time: now,
+				jop_updated_time: now,
+				content: requestBody.toString(),
+			};
+			return JSON.stringify(output);
+		});
+
+		// {
+		// 	'locks/exclusive_cli_cc75ed109c0c40d5ac8707d222fe33bc.json': {
+		// 	  item: {
+		// 		name: 'locks/exclusive_cli_cc75ed109c0c40d5ac8707d222fe33bc.json',
+		// 		updated_time: 1635709007725,
+		// 		id: 'v9Yv5WSxAKF75ZW0dnv8nOQ8hzg5rUz1'
+		// 	  },
+		// 	  error: null
+		// 	}
+		//   }
+
+		return {
+			handled: true,
+			response: {
+				[itemName]: {
+					item: {
+						name: itemName,
+						updated_time: now,
+						id: null,
+					},
+					error: null,
+				},
+			},
+		};
+	}
+
+	// DELETE /api/items/root:/locks/exclusive_cli_12cb74fa9de644958b2ccbc772cb4e29.json:
+
+	if (ctx.method === 'DELETE') {
+		const itemName = models.item().pathToName(path.id);
+
+		await models.keyValue().readThenWrite(userKey, async (value: Value) => {
+			const output = value ? JSON.parse(value as string) : {};
+			delete output[itemName];
+			return JSON.stringify(output);
+		});
+
+		return {
+			handled: true,
+			response: null,
+		};
+	}
+
+	// GET /api/items/root:/locks/*:/children
+
+	if (ctx.method === 'GET' && path.id === 'root:/locks/*:') {
+		const result = await models.keyValue().value<string>(userKey);
+		const obj: Record<string, Item> = result ? JSON.parse(result) : {};
+
+		const items: Item[] = [];
+		for (const name of Object.keys(obj)) {
+			items.push(obj[name]);
+		}
+
+		const page: PaginatedItems = {
+			has_more: false,
+			items,
+		};
+
+		return {
+			handled: true,
+			response: page,
+		};
+	}
+
+	throw new ErrorBadRequest(`Unhandled lock path: ${path.id}`);
+};
 
 export async function putItemContents(path: SubPath, ctx: AppContext, isBatch: boolean) {
 	if (!ctx.joplin.owner.can_upload) throw new ErrorForbidden('Uploading content is disabled');
@@ -41,6 +141,9 @@ export async function putItemContents(path: SubPath, ctx: AppContext, isBatch: b
 
 		try {
 			const buffer = filePath ? await fs.readFile(filePath) : Buffer.alloc(0);
+
+			const lockResult = await lockHandler(path, ctx, buffer);
+			if (lockResult.handled) return lockResult.response;
 
 			// This end point can optionally set the associated jop_share_id field. It
 			// is only useful when uploading resource blob (under .resource folder)
@@ -104,8 +207,13 @@ router.del('api/items/:id', async (path: SubPath, ctx: AppContext) => {
 			if (ctx.joplin.env !== 'dev') throw new ErrorMethodNotAllowed('Deleting the root is not allowed');
 			await ctx.joplin.models.item().deleteAll(ctx.joplin.owner.id);
 		} else {
+			// const item = await itemFromPath(ctx.joplin.owner.id, ctx.joplin.models.item(), path);
+			// await ctx.joplin.models.item().checkIfAllowed(ctx.joplin.owner, AclAction.Delete, item);
+
+			const lockResult = await lockHandler(path, ctx);
+			if (lockResult.handled) return lockResult.response;
+
 			const item = await itemFromPath(ctx.joplin.owner.id, ctx.joplin.models.item(), path);
-			await ctx.joplin.models.item().checkIfAllowed(ctx.joplin.owner, AclAction.Delete, item);
 			await ctx.joplin.models.item().deleteForUser(ctx.joplin.owner.id, item);
 		}
 	} catch (error) {
@@ -137,6 +245,9 @@ router.get('api/items/:id/delta', async (_path: SubPath, ctx: AppContext) => {
 });
 
 router.get('api/items/:id/children', async (path: SubPath, ctx: AppContext) => {
+	const lockResult = await lockHandler(path, ctx);
+	if (lockResult.handled) return lockResult.response;
+
 	const itemModel = ctx.joplin.models.item();
 	const parentName = itemModel.pathToName(path.id);
 	const result = await itemModel.children(ctx.joplin.owner.id, parentName, requestPagination(ctx.query));
